+++ app/routes.js
@@ -5,10 +5,7 @@ import Relay from 'react-relay/classic';
 import { Route, IndexRoute, IndexRedirect } from 'react-router';
 import ContainerDimensions from 'react-container-dimensions';
 
-import omitBy from 'lodash/omitBy';
-import isNil from 'lodash/isNil';
-
-import moment from 'moment';
+import { preparePlanParams } from './util/planParamUtil';
 
 // React pages
 import IndexPage from './component/IndexPage';
@@ -27,9 +24,6 @@ import {
   PREFIX_STOPS,
   PREFIX_ITINERARY_SUMMARY,
 } from './util/path';
-
-// Localstorage data
-import { getCustomizedSettings } from './store/localStorage';
 
 const ComponentLoading404Renderer = {
   /* eslint-disable react/prop-types */
@@ -139,39 +133,6 @@ function getIntermediatePlaces(intermediatePlaces) {
   return [];
 }
 
-function getSettings() {
-  const custSettings = getCustomizedSettings();
-
-  return {
-    walkSpeed: custSettings.walkSpeed
-      ? Number(custSettings.walkSpeed)
-      : undefined,
-    walkReluctance: custSettings.walkReluctance
-      ? Number(custSettings.walkReluctance)
-      : undefined,
-    walkBoardCost: custSettings.walkBoardCost
-      ? Number(custSettings.walkBoardCost)
-      : undefined,
-    modes: custSettings.modes
-      ? custSettings.modes
-          .toString()
-          .split(',')
-          .map(mode => (mode === 'CITYBIKE' ? 'BICYCLE_RENT' : mode))
-          .sort()
-          .join(',')
-      : undefined,
-    minTransferTime: custSettings.minTransferTime
-      ? Number(custSettings.minTransferTime)
-      : undefined,
-    accessibilityOption: custSettings.accessibilityOption
-      ? custSettings.accessibilityOption
-      : undefined,
-    ticketTypes: custSettings.ticketTypes
-      ? custSettings.ticketTypes
-      : undefined,
-  };
-}
-
 function setTicketTypes(ticketType, settingsTicketType) {
   if (ticketType !== undefined && ticketType !== 'none') {
     return ticketType;
@@ -186,81 +147,12 @@ function setTicketTypes(ticketType, settingsTicketType) {
 }
 
 export default config => {
-  const preparePlanParams = (
-    { from, to },
-    {
-      location: {
-        query: {
-          intermediatePlaces,
-          numItineraries,
-          time,
-          arriveBy,
-          walkReluctance,
-          walkSpeed,
-          walkBoardCost,
-          minTransferTime,
-          modes,
-          accessibilityOption,
-          ticketTypes,
-        },
-      },
-    },
-  ) => {
-    const settings = getSettings();
-    return {
-      ...omitBy(
-        {
-          fromPlace: from,
-          toPlace: to,
-          from: otpToLocation(from),
-          to: otpToLocation(to),
-          intermediatePlaces: getIntermediatePlaces(intermediatePlaces),
-          numItineraries: numItineraries ? Number(numItineraries) : undefined,
-          modes: modes
-            ? modes
-                .split(',')
-                .map(mode => (mode === 'CITYBIKE' ? 'BICYCLE_RENT' : mode))
-                .sort()
-                .join(',')
-            : settings.modes,
-          date: time ? moment(time * 1000).format('YYYY-MM-DD') : undefined,
-          time: time ? moment(time * 1000).format('HH:mm:ss') : undefined,
-          walkReluctance: walkReluctance
-            ? Number(walkReluctance)
-            : settings.walkReluctance,
-          walkBoardCost: walkBoardCost
-            ? Number(walkBoardCost)
-            : settings.walkBoardCost,
-          minTransferTime: minTransferTime
-            ? Number(minTransferTime)
-            : settings.minTransferTime,
-          walkSpeed: walkSpeed ? Number(walkSpeed) : settings.walkSpeed,
-          arriveBy: arriveBy ? arriveBy === 'true' : undefined,
-          maxWalkDistance:
-            typeof modes === 'undefined' ||
-            (typeof modes === 'string' && !modes.split(',').includes('BICYCLE'))
-              ? config.maxWalkDistance
-              : config.maxBikingDistance,
-          wheelchair:
-            accessibilityOption === '1'
-              ? true
-              : settings.accessibilityOption === '1',
-          preferred: { agencies: config.preferredAgency || '' },
-          disableRemainingWeightHeuristic:
-            modes && modes.split(',').includes('CITYBIKE'),
-        },
-        isNil,
-      ),
-      ticketTypes: setTicketTypes(ticketTypes, settings.ticketTypes),
-    };
-  };
-
   const SummaryPageWrapper = ({ props, routerProps, element }) =>
     props
       ? React.cloneElement(element, props)
       : React.cloneElement(element, {
           ...routerProps,
-          ...preparePlanParams(routerProps.params, routerProps),
+          ...preparePlanParams(config)(routerProps.params, routerProps),
           plan: { plan: {} },
           loading: true,
         });
@@ -571,7 +463,7 @@ export default config => {
           );
         }}
         queries={{ content: planQueries }}
-        prepareParams={preparePlanParams}
+        prepareParams={preparePlanParams(config)}
         render={{ content: SummaryPageWrapper }}
       >
         <Route
